requirements.txt
paho-mqtt
solax
PyYAML
